cli: move add_to_git_exclude() helper to commands.git module
I'm going to add a call site to the git module, and I think it's better to
host the helper function there instead of importing from the super module.
